test/imgtestlib: break up imgtestlib into multiple submodules
The module got too big for one file.

This commit only moves code around and adjusts imports.  Functions are
grouped into submodules by logical functionality.  Groups (submodule
boundaries) were chosen to avoid circular imports.

The .core module could still use some splitting or tidying up, but the
current state is already an improvement.

Importing everything into __init__.py might be unnecessary—we only
really need to import the functions that are used externally—but it's
not very important to reduce the API surface of this internal testing
library.
