+++ test/scripts/imgtestlib/__init__.py
@@ -1,1 +1,7 @@
-from .imgtestlib import *
+from .boot import *
+from .build import *
+from .cache import *
+from .core import *
+from .gitlab import *
+from .run import *
+from .testenv import *
