+++ test/scripts/imgtestlib/cache.py
@@ -1,0 +1,112 @@
+import os
+import subprocess as sp
+from datetime import datetime
+from typing import List, Optional
+
+from .run import runcmd_nc
+from .testenv import get_host_distro, get_osbuild_commit
+
+S3_BUCKET = "s3://" + os.environ.get("AWS_BUCKET", "images-ci-cache")
+S3_PREFIX = "images/builds"
+
+
+# pylint: disable=too-many-arguments,too-many-positional-arguments
+def dl_build_cache(
+        destination, distro: Optional[str] = None, arch: Optional[str] = None, osbuild_ref: Optional[str] = None,
+        runner_distro: Optional[str] = None, manifest_id: Optional[str] = None,
+        include_only: Optional[List[str]] = None):
+    """
+    Downloads image build cache files from the s3 bucket.
+
+    If 'include' is not specified, all files are downloaded. Otherwise, all files will be excluded and the items
+    in the 'include' list will be passed as '--include' arguments to the 'aws s3 sync' command.
+    """
+    s3url = gen_build_info_s3_dir_path(distro, arch, manifest_id, osbuild_ref, runner_distro)
+    dl_what = "all files" if include_only is None else "only " + ', '.join(f"'{i}'" for i in include_only)
+    print(f"⬇️ Downloading {dl_what} from {s3url}")
+
+    # --no-progress wont show progress but will print file list
+    cmd = ["aws", "s3", "sync", "--no-progress"]
+    if include_only:
+        cmd.extend(["--exclude=*"])
+        for i in include_only:
+            cmd.extend([f"--include={i}"])
+    cmd.extend([s3url, destination])
+
+    job = sp.run(cmd, capture_output=True, check=False)
+    ok = job.returncode == 0
+    if not ok:
+        print(f"⚠️ Failed to sync contents of {s3url}:")
+        print(job.stdout.decode())
+        print(job.stderr.decode())
+    return job.stdout.decode(), ok
+
+
+def dl_build_info(destination, distro=None, arch=None, osbuild_ref=None, runner_distro=None):
+    """
+    Downloads all the configs from the s3 bucket.
+    """
+    # only download info.json and bib-* files, otherwise we get manifests and whole images
+    include = ["*/info.json", "*/bib-*"]
+    return dl_build_cache(destination, distro, arch, osbuild_ref, runner_distro, include_only=include)
+
+
+def gen_build_info_dir_path_prefix(distro=None, arch=None, manifest_id=None, osbuild_ref=None, runner_distro=None):
+    """
+    Generates the relative path prefix for the location where build info and artifacts will be stored for a specific
+    build. This is a simple concatenation of the components, but ensures that paths are consistent. The caller is
+    responsible for prepending the location root to the generated path.
+
+    If no 'osbuild_ref' is specified, the value returned by get_osbuild_commit() for the 'runner_distro' will be used.
+    if no 'runner_distro' is specified, the value returned by get_host_distro() will be used.
+
+    A fully specified path is returned if all of the 'distro', 'arch' and 'manifest_id' parameters are specified,
+    otherwise a partial path is returned. Partial path may be useful for working with a superset of build infos.
+    For a more specific path to be generated when specifying any of the optional parameters, the caller must specify
+    all of the previous parameters. For example, if 'arch' is specified, 'distro' must also be specified for 'arch' to
+    be included in the path.
+
+    The returned path always has a trailing separator at the end to signal that it is a directory.
+    """
+    if runner_distro is None:
+        runner_distro = get_host_distro()
+    if osbuild_ref is None:
+        osbuild_ref = get_osbuild_commit(runner_distro)
+
+    path = os.path.join(f"osbuild-ref-{osbuild_ref}", f"runner-{runner_distro}")
+    for p in (distro, arch, f"manifest-id-{manifest_id}" if manifest_id else None):
+        if p is None:
+            return path + "/"
+        path = os.path.join(path, p)
+    return path + "/"
+
+
+def gen_build_info_s3_dir_path(distro=None, arch=None, manifest_id=None, osbuild_ref=None, runner_distro=None):
+    """
+    Generates the s3 URL for the location where build info and artifacts will be stored for a specific
+    one or more builds, depending on the parameters specified.
+
+    A fully specified path is returned if all parameters are specified, otherwise a partial path is returned.
+    This function basically just prepends the S3_BUCKET and S3_PREFIX to the path generated by
+    gen_build_info_dir_path_prefix().
+    """
+    return os.path.join(
+        S3_BUCKET,
+        S3_PREFIX,
+        gen_build_info_dir_path_prefix(distro, arch, manifest_id, osbuild_ref, runner_distro),
+    )
+
+
+def touch_s3(distro, arch, manifest_id, osbuild_ref=None, runner_distro=None):
+    """
+    Update the timestamps of a path in S3 by adding a metadata field to each file recursively. This can be used to
+    "freshen up" relevant files in the build cache so that images that are still current but haven't been updated in a
+    while don't get garbage collected.
+    """
+    # NOTE: we can make this async - updating the objects can take a few seconds and can be done in parallel
+    s3url = gen_build_info_s3_dir_path(distro, arch, manifest_id, osbuild_ref, runner_distro)
+    # the exact key and value don't matter, but let's add the current datetime to make it a bit more meaningful
+    now = str(datetime.now())
+    print(f"⌚ Updating timestamps for {s3url} ({now})")
+    cmd = ["aws", "s3", "cp", "--recursive", "--metadata", f"touched={now}", s3url, s3url]
+    runcmd_nc(cmd)
